+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `equality` operator overload
+- `gc_clamp` property
 ## [0.1] - 2024-11-27
 ### Added
 - `MeltingTemperature` enum